backend/requirements.txt
flask
pytesseract
pillow
flask-cors
